Create separate modules for BOM and HTML functions
