+++ src/wireviz/wv_helper.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from wireviz import wv_colors
 from typing import List
 import re
 
@@ -31,58 +30,6 @@ def awg_equiv(mm2):
 
 def mm2_equiv(awg):
     return mm2_equiv_table.get(str(awg), 'Unknown')
-
-def nested_html_table(rows):
-    # input: list, each item may be scalar or list
-    # output: a parent table with one child table per parent item that is list, and one cell per parent item that is scalar
-    # purpose: create the appearance of one table, where cell widths are independent between rows
-    # attributes in any leading <tdX> inside a list are injected into to the preceeding <td> tag
-    html = []
-    html.append('<table border="0" cellspacing="0" cellpadding="0">')
-    for row in rows:
-        if isinstance(row, List):
-            if len(row) > 0 and any(row):
-                html.append(' <tr><td>')
-                html.append('  <table border="0" cellspacing="0" cellpadding="3" cellborder="1"><tr>')
-                for cell in row:
-                    if cell is not None:
-                        # Inject attributes to the preceeding <td> tag where needed
-                        html.append(f'   <td balign="left">{cell}</td>'.replace('><tdX', ''))
-                html.append('  </tr></table>')
-                html.append(' </td></tr>')
-        elif row is not None:
-            html.append(' <tr><td>')
-            html.append(f'  {row}')
-            html.append(' </td></tr>')
-    html.append('</table>')
-    return html
-
-def html_colorbar(color):
-    return f'<tdX bgcolor="{wv_colors.translate_color(color, "HEX")}" width="4">' if color else None
-
-def html_image(image):
-    if not image:
-        return None
-    # The leading attributes belong to the preceeding tag. See where used below.
-    html = f'{html_size_attr(image)}><img scale="{image.scale}" src="{image.src}"/>'
-    if image.fixedsize:
-        # Close the preceeding tag and enclose the image cell in a table without
-        # borders to avoid narrow borders when the fixed width < the node width.
-        html = f'''>
-    <table border="0" cellspacing="0" cellborder="0"><tr>
-     <td{html}</td>
-    </tr></table>
-   '''
-    return f'''<tdX{' sides="TLR"' if image.caption else ''}{html}'''
-
-def html_caption(image):
-    return f'<tdX sides="BLR">{html_line_breaks(image.caption)}' if image and image.caption else None
-
-def html_size_attr(image):
-    # Return Graphviz HTML attributes to specify minimum or fixed size of a TABLE or TD object
-    return ((f' width="{image.width}"'   if image.width else '')
-        +   (f' height="{image.height}"' if image.height else '')
-        +   ( ' fixedsize="true"'        if image.fixedsize else '')) if image else ''
 
 
 def expand(yaml_data):
@@ -140,18 +87,14 @@ def tuplelist2tsv(inp, header=None):
         output = output + '\t'.join(str(remove_links(item)) for item in row) + '\n'
     return output
 
-# Return the value indexed if it is a list, or simply the value otherwise.
-def index_if_list(value, index):
-    return value[index] if isinstance(value, list) else value
 
 def remove_links(inp):
     return re.sub(r'<[aA] [^>]*>([^<]*)</[aA]>', r'\1', inp) if isinstance(inp, str) else inp
 
-def html_line_breaks(inp):
-    return remove_links(inp).replace('\n', '<br />') if isinstance(inp, str) else inp
 
 def clean_whitespace(inp):
     return ' '.join(inp.split()).replace(' ,', ',') if isinstance(inp, str) else inp
+
 
 def open_file_read(filename):
     # TODO: Intelligently determine encoding
@@ -162,7 +105,6 @@ def open_file_write(filename):
 
 def open_file_append(filename):
     return open(filename, 'a', encoding='UTF-8')
-
 
 def aspect_ratio(image_src):
     try:
@@ -175,32 +117,3 @@ def aspect_ratio(image_src):
     except Exception as error:
         print(f'aspect_ratio(): {type(error).__name__}: {error}')
     return 1 # Assume 1:1 when unable to read actual image size
-
-
-def manufacturer_info_field(manufacturer, mpn):
-    if manufacturer or mpn:
-        return f'{manufacturer if manufacturer else "MPN"}{": " + str(mpn) if mpn else ""}'
-    else:
-        return None
-
-def component_table_entry(type, qty, unit=None, pn=None, manufacturer=None, mpn=None):
-    output = f'{qty}'
-    if unit:
-        output += f' {unit}'
-    output += f' x {type}'
-    # print an extra line with part and manufacturer information if provided
-    manufacturer_str = manufacturer_info_field(manufacturer, mpn)
-    if pn or manufacturer_str:
-        output += '<br/>'
-        if pn:
-            output += f'P/N: {pn}'
-            if manufacturer_str:
-                output += ', '
-        if manufacturer_str:
-            output += manufacturer_str
-    output = html_line_breaks(output)
-    # format the above output as left aligned text in a single visible cell
-    # indent is set to two to match the indent in the generated html table
-    return f'''<table border="0" cellspacing="0" cellpadding="3" cellborder="1"><tr>
-   <td align="left" balign="left">{output}</td>
-  </tr></table>'''
